Moves auditing tasks from test to gulp audit task
